requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.6.1
anthropic==0.8.1
requests==2.31.0
email-validator==2.1.0
reportlab==4.0.9
beautifulsoup4==4.12.2
markdown==3.5.2
pypdf==3.17.4
pdf2image==1.17.0
pytesseract==0.3.10
wtforms==3.1.1
itsdangerous==2.1.2
oauthlib==3.2.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytest==7.4.3
pytest-mock==3.12.0
flask-testing==0.8.1
flask-migrate==4.0.5
alembic==1.13.1
celery==5.3.4
redis==5.0.1
boto3==1.34.0
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0